Add ARMv8-based CRC accelerated checksum
```
...
In Armv8-A, this is an OPTIONAL instruction, and in Armv8.1 it is mandatory for all implementations to implement it.
...
```
https://developer.arm.com/documentation/ddi0597/2020-12/Base-Instructions/CRC32--CRC32-

Uses the `crc32{b,h,w,d}` and `crc32c{b,h,w,d}`instructions.  Dispatch
is determined at compile-time using the `__ARM_FEATURE_CRC32`
preprocessor.

Tested on an M2 Mac Mini, it's not all that faster, might be something
I'm doing wrong?

Before:
```
./qCheck-generic -c   26.95s user 3.82s system 104% cpu 29.345 total
```

After:
```
./qCheck-crc32 -c   26.85s user 3.69s system 104% cpu 29.204 total
```